Create Callback_2.sv
The callbacks are used to alter the behavior of the component without modifying its code. The verification engineer provides a set of hook methods that helps to customize the behavior depending on the requirement. A simple example of callbacks can be the pre_randomize and post_randomize methods before and after the built-in randomize method call.